Added dbapi2 connection logic to schema.
We will get data with raw SQL. This is done as a time saving measure. We're using sql-alchemy to save some hassle in managing the schema, but we are not using its ability to abstract away the database since sqlite will do everything we need, and we can move much more quickly writing raw queries.

Also modified where the database gets created. We make it at the base of the project directory now. This is purely a convenience thing.